+++ src/itables/dash.py
@@ -1,13 +1,26 @@
-from itables_for_dash import (
+from itables import __version__
+from itables_for_dash.properties import (
     ITABLE_PROPERTIES,
-    ITable,
-    ITableComponent,
     ITableOutputs,
-    __version__,
     get_itable_component_kwargs,
-    itables_for_dash_is_available,
     updated_itable_outputs,
 )
+
+try:
+    from itables_for_dash import ITable, ITableComponent  # type: ignore
+except ImportError as e:
+    import_error = e
+
+    class ITableComponent:
+        def __init__(self, **kwargs):
+            raise import_error
+
+    class ITable(ITableComponent):
+        pass
+
+    itables_for_dash_is_available = False
+else:
+    itables_for_dash_is_available = True
 
 __all__ = [
     "ITable",
@@ -16,6 +29,5 @@ __all__ = [
     "get_itable_component_kwargs",
     "ITableOutputs",
     "updated_itable_outputs",
-    "itables_for_dash_is_available",
     "__version__",
 ]