+++ src/itables/downsample.py
@@ -31,7 +31,7 @@ def as_nbytes(mem):
     return int(float(mem))
 
 
-def downsample(df, max_rows=0, max_columns=0, max_bytes=0):
+def downsample(df, max_rows: int = 0, max_columns: int = 0, max_bytes: int | str = 0):
     """Return a subset of the dataframe that fits the limits"""
     org_rows, org_columns, org_bytes = len(df), len(df.columns), nbytes(df)
     max_bytes_numeric = as_nbytes(max_bytes)
@@ -113,16 +113,16 @@ def _downsample(df, max_rows=0, max_columns=0, max_bytes=0, target_aspect_ratio=
         first_half = max_columns - second_half
         assert first_half >= second_half
         if second_half:
-            try:
+            if isinstance(df, pd.DataFrame):
                 df = pd.concat(
                     (df.iloc[:, :first_half], df.iloc[:, -second_half:]), axis=1
                 )
-            except AttributeError:
+            else:
                 df = df[df.columns[:first_half]].hstack(df[df.columns[-second_half:]])
         else:
-            try:
+            if isinstance(df, pd.DataFrame):
                 df = df.iloc[:, :first_half]
-            except AttributeError:
+            else:
                 df = df[df.columns[:first_half]]
 
     df_nbytes = nbytes(df)